+++ src/api/annotations/$Description.ts
@@ -1,6 +1,6 @@
 import { $Push } from 'jest-metadata';
 
-import { DESCRIPTION } from '../constants';
+import { DESCRIPTION } from '../../constants';
 
 export const $Description = (description: string) =>
   $Push(DESCRIPTION, description);
