+++ src/api/annotations/$DescriptionHtml.ts
@@ -1,6 +1,6 @@
 import { $Push } from 'jest-metadata';
 
-import { DESCRIPTION_HTML } from '../constants';
+import { DESCRIPTION_HTML } from '../../constants';
 
 export const $DescriptionHtml = (descriptionHtml: string) =>
   $Push(DESCRIPTION_HTML, descriptionHtml);
