+++ src/api/runtime/IAllureRuntime.ts
@@ -6,8 +6,9 @@ import type {
   Status,
   StatusDetails,
 } from 'jest-allure2-reporter';
+import type { Metadata } from 'jest-metadata';
 
-import type { Function_, MaybePromise } from '../utils/types';
+// import type { Function_, MaybePromise } from '../utils/types';
 
 export interface IAllureRuntime {
   /**
@@ -15,6 +16,12 @@ export interface IAllureRuntime {
    * Useful when your artifacts are delayed and are created asynchronously.
    */
   $bind(options?: AllureRuntimeBindOptions): IAllureRuntime;
+
+  /**
+   * Attach a runtime plugin using a callback.
+   * The callback will be called with the runtime plugin context.
+   */
+  $plug(callback: AllureRuntimePluginCallback): void;
 
   description(value: string): void;
 
@@ -32,49 +39,49 @@ export interface IAllureRuntime {
 
   parameters(parameters: Record<string, unknown>): void;
 
-  attachment<T extends AttachmentContent>(
-    name: string,
-    content: MaybePromise<T>,
-    mimeType?: string,
-  ): typeof content;
-
-  createAttachment<
-    T extends AttachmentContent,
-    F extends Function_<MaybePromise<T>>,
-  >(
-    function_: F,
-    name: string,
-  ): F;
-  createAttachment<
-    T extends AttachmentContent,
-    F extends Function_<MaybePromise<T>>,
-  >(
-    function_: F,
-    options: ContentAttachmentOptions,
-  ): typeof function_;
-
-  fileAttachment(filePath: string, name?: string): string;
-  fileAttachment(filePath: string, options?: FileAttachmentOptions): string;
-  fileAttachment(
-    filePathPromise: Promise<string>,
-    name?: string,
-  ): Promise<string>;
-  fileAttachment(
-    filePathPromise: Promise<string>,
-    options?: FileAttachmentOptions,
-  ): Promise<string>;
-
-  createFileAttachment<F extends Function_<MaybePromise<string>>>(
-    function_: F,
-  ): F;
-  createFileAttachment<F extends Function_<MaybePromise<string>>>(
-    function_: F,
-    name: string,
-  ): F;
-  createFileAttachment<F extends Function_<MaybePromise<string>>>(
-    function_: F,
-    options: FileAttachmentOptions,
-  ): F;
+  // attachment<T extends AttachmentContent>(
+  //   name: string,
+  //   content: MaybePromise<T>,
+  //   mimeType?: string,
+  // ): typeof content;
+  //
+  // createAttachment<
+  //   T extends AttachmentContent,
+  //   F extends Function_<MaybePromise<T>>,
+  // >(
+  //   function_: F,
+  //   name: string,
+  // ): F;
+  // createAttachment<
+  //   T extends AttachmentContent,
+  //   F extends Function_<MaybePromise<T>>,
+  // >(
+  //   function_: F,
+  //   options: ContentAttachmentOptions,
+  // ): typeof function_;
+  //
+  // fileAttachment(filePath: string, name?: string): string;
+  // fileAttachment(filePath: string, options?: FileAttachmentOptions): string;
+  // fileAttachment(
+  //   filePathPromise: Promise<string>,
+  //   name?: string,
+  // ): Promise<string>;
+  // fileAttachment(
+  //   filePathPromise: Promise<string>,
+  //   options?: FileAttachmentOptions,
+  // ): Promise<string>;
+  //
+  // createFileAttachment<F extends Function_<MaybePromise<string>>>(
+  //   function_: F,
+  // ): F;
+  // createFileAttachment<F extends Function_<MaybePromise<string>>>(
+  //   function_: F,
+  //   name: string,
+  // ): F;
+  // createFileAttachment<F extends Function_<MaybePromise<string>>>(
+  //   function_: F,
+  //   options: FileAttachmentOptions,
+  // ): F;
 
   createStep<F extends Function>(name: string, function_: F): F;
   createStep<F extends Function>(
@@ -84,6 +91,17 @@ export interface IAllureRuntime {
   ): F;
 
   step<T>(name: string, function_: () => T): T;
+}
+
+export type AllureRuntimePluginCallback = (
+  context: AllureRuntimePluginContext,
+) => void;
+
+export interface AllureRuntimePluginContext {
+  contentHandlers: Record<string, ContentAttachmentHandler>;
+  fileHandlers: Record<string, FileAttachmentHandler>;
+  getMetadata: () => Metadata;
+  runtime: IAllureRuntime;
 }
 
 export type FileAttachmentOptions = {
