+++ src/api/runtime/proxies/index.ts
@@ -1,0 +1,2 @@
+export * from './AllureTestItemMetadataProxy';
+export * from './AllureTestCaseMetadataProxy';
